Introduce new board feature flag
