+++ packages/twenty-front/src/App.tsx
@@ -1,11 +1,10 @@
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 
-import { RecordShowPage } from '@/object-record/components/RecordShowPage';
-import { RecordTablePage } from '@/object-record/components/RecordTablePage';
 import { AppPath } from '@/types/AppPath';
 import { SettingsPath } from '@/types/SettingsPath';
 import { DefaultLayout } from '@/ui/layout/page/DefaultLayout';
 import { PageTitle } from '@/ui/utilities/page-title/PageTitle';
+import { useIsFeatureEnabled } from '@/workspace/hooks/useIsFeatureEnabled';
 import { CommandMenuEffect } from '~/effect-components/CommandMenuEffect';
 import { GotoHotkeysEffect } from '~/effect-components/GotoHotkeysEffect';
 import { useDefaultHomePagePath } from '~/hooks/useDefaultHomePagePath';
@@ -16,6 +15,8 @@ import { SignInUp } from '~/pages/auth/SignInUp';
 import { VerifyEffect } from '~/pages/auth/VerifyEffect';
 import { ImpersonateEffect } from '~/pages/impersonate/ImpersonateEffect';
 import { NotFound } from '~/pages/not-found/NotFound';
+import { RecordIndexPage } from '~/pages/object-record/RecordIndexPage';
+import { RecordShowPage } from '~/pages/object-record/RecordShowPage';
 import { Opportunities } from '~/pages/opportunities/Opportunities';
 import { SettingsAccounts } from '~/pages/settings/accounts/SettingsAccounts';
 import { SettingsAccountsEmails } from '~/pages/settings/accounts/SettingsAccountsEmails';
@@ -43,6 +44,9 @@ export const App = () => {
   const { defaultHomePagePath } = useDefaultHomePagePath();
 
   const pageTitle = getPageTitleFromPath(pathname);
+  const isNewRecordBoardEnabled = useIsFeatureEnabled(
+    'IS_NEW_RECORD_BOARD_ENABLED',
+  );
 
   return (
     <>
@@ -62,8 +66,13 @@ export const App = () => {
           <Route path={AppPath.TasksPage} element={<Tasks />} />
           <Route path={AppPath.Impersonate} element={<ImpersonateEffect />} />
 
-          <Route path={AppPath.OpportunitiesPage} element={<Opportunities />} />
-          <Route path={AppPath.RecordTablePage} element={<RecordTablePage />} />
+          {!isNewRecordBoardEnabled && (
+            <Route
+              path={AppPath.OpportunitiesPage}
+              element={<Opportunities />}
+            />
+          )}
+          <Route path={AppPath.RecordIndexPage} element={<RecordIndexPage />} />
           <Route path={AppPath.RecordShowPage} element={<RecordShowPage />} />
 
           <Route
