+++ packages/twenty-front/src/modules/object-record/record-index/components/RecordTableEffect.tsx
