+++ packages/twenty-front/src/modules/types/AppPath.ts
@@ -14,8 +14,8 @@ export enum AppPath {
   Index = '/',
   TasksPage = '/tasks',
   OpportunitiesPage = '/objects/opportunities',
-  RecordTablePage = '/objects/:objectNamePlural',
 
+  RecordIndexPage = '/objects/:objectNamePlural',
   RecordShowPage = '/object/:objectNameSingular/:objectRecordId',
 
   SettingsCatchAll = `/settings/*`,
