+++ packages/twenty-front/src/modules/workspace/types/FeatureFlagKey.ts
@@ -1,4 +1,5 @@
 export type FeatureFlagKey =
   | 'IS_MESSAGING_ENABLED'
   | 'IS_QUICK_ACTIONS_ENABLED'
-  | 'IS_RATING_FIELD_TYPE_ENABLED';
+  | 'IS_RATING_FIELD_TYPE_ENABLED'
+  | 'IS_NEW_RECORD_BOARD_ENABLED';
