+++ packages/twenty-front/src/pages/object-record/RecordShowPage.tsx
@@ -1,0 +1,113 @@
+import { useEffect } from 'react';
+import { useParams } from 'react-router-dom';
+import { useSetRecoilState } from 'recoil';
+
+import { useFavorites } from '@/favorites/hooks/useFavorites';
+import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
+import { entityFieldsFamilyState } from '@/object-record/field/states/entityFieldsFamilyState';
+import { RecordShowContainer } from '@/object-record/record-show/components/RecordShowContainer';
+import { IconBuildingSkyscraper } from '@/ui/display/icon';
+import { PageBody } from '@/ui/layout/page/PageBody';
+import { PageContainer } from '@/ui/layout/page/PageContainer';
+import { PageFavoriteButton } from '@/ui/layout/page/PageFavoriteButton';
+import { PageHeader } from '@/ui/layout/page/PageHeader';
+import { ShowPageAddButton } from '@/ui/layout/show-page/components/ShowPageAddButton';
+import { ShowPageMoreButton } from '@/ui/layout/show-page/components/ShowPageMoreButton';
+import { PageTitle } from '@/ui/utilities/page-title/PageTitle';
+import { isDefined } from '~/utils/isDefined';
+
+import { useFindOneRecord } from '../../modules/object-record/hooks/useFindOneRecord';
+
+export const RecordShowPage = () => {
+  const { objectNameSingular, objectRecordId } = useParams<{
+    objectNameSingular: string;
+    objectRecordId: string;
+  }>();
+
+  if (!objectNameSingular) {
+    throw new Error(`Object name is not defined`);
+  }
+
+  if (!objectRecordId) {
+    throw new Error(`Record id is not defined`);
+  }
+
+  const { objectMetadataItem } = useObjectMetadataItem({
+    objectNameSingular,
+  });
+
+  const { favorites, createFavorite, deleteFavorite } = useFavorites();
+
+  const setEntityFields = useSetRecoilState(
+    entityFieldsFamilyState(objectRecordId ?? ''),
+  );
+
+  const { record } = useFindOneRecord({
+    objectRecordId,
+    objectNameSingular,
+  });
+
+  useEffect(() => {
+    if (!record) return;
+    setEntityFields(record);
+  }, [record, setEntityFields]);
+
+  const correspondingFavorite = favorites.find(
+    (favorite) => favorite.recordId === objectRecordId,
+  );
+
+  const isFavorite = isDefined(correspondingFavorite);
+
+  const handleFavoriteButtonClick = async () => {
+    if (!objectNameSingular || !record) return;
+
+    if (isFavorite && record) {
+      deleteFavorite(correspondingFavorite.id);
+    } else {
+      createFavorite(record, objectNameSingular);
+    }
+  };
+
+  const pageName =
+    objectNameSingular === 'person'
+      ? record?.name.firstName + ' ' + record?.name.lastName
+      : record?.name;
+
+  return (
+    <PageContainer>
+      <PageTitle title={pageName} />
+      <PageHeader
+        title={pageName ?? ''}
+        hasBackButton
+        Icon={IconBuildingSkyscraper}
+      >
+        {record && (
+          <>
+            <PageFavoriteButton
+              isFavorite={isFavorite}
+              onClick={handleFavoriteButtonClick}
+            />
+            <ShowPageAddButton
+              key="add"
+              entity={{
+                id: record.id,
+                targetObjectNameSingular: objectMetadataItem?.nameSingular,
+              }}
+            />
+            <ShowPageMoreButton
+              key="more"
+              recordId={record.id}
+              objectNameSingular={objectNameSingular}
+            />
+          </>
+        )}
+      </PageHeader>
+      <PageBody>
+        <RecordShowContainer
+          objectNameSingular={objectNameSingular}
+          objectRecordId={objectRecordId}
+        />
+      </PageBody>
+    </PageContainer>
+  );
+};
