+++ packages/twenty-server/src/core/feature-flag/feature-flag.entity.ts
@@ -17,6 +17,7 @@ export enum FeatureFlagKeys {
   IsMessagingEnabled = 'IS_MESSAGING_ENABLED',
   IsRatingFieldTypeEnabled = 'IS_RATING_FIELD_TYPE_ENABLED',
   IsWorkspaceCleanable = 'IS_WORKSPACE_CLEANABLE',
+  IsNewRecordBoardEnabled = 'IS_NEW_RECORD_BOARD_ENABLED',
 }
 
 @Entity({ name: 'featureFlag', schema: 'core' })
